Moves logic and functions of avatar update to a new child component. All tests passing.
