+++ src/index.js
@@ -11,7 +11,7 @@ import axios from 'axios'
 const development = 'http://localhost:3007'
 const production = process.env.REACT_APP_API_ENDPOINT
 
-axios.defaults.baseURL = development
+axios.defaults.baseURL = production
 
 const store = configureStore()
 verifyCredentials(store)
