+++ src/reduxTokenAuthConfig.js
@@ -4,7 +4,7 @@ const production = `${process.env.REACT_APP_API_ENDPOINT}/api/v1/auth`
 const development = 'http://localhost:3007/api/v1/auth'
 
 const config = {
-  authUrl: development,
+  authUrl: production,
   userAttributes: {
     id: 'id',
     uid: 'uid',
